+++ CavrnusSdk/Runtime/Scripts/Setup/CavrnusSpatialConnector.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
+using CavrnusCore;
 using UnityEngine;
 using CavrnusSdk.API;
+using CavrnusSdk.PropertySynchronizers;
 
 namespace CavrnusSdk.Setup
 {
@@ -49,8 +51,7 @@ namespace CavrnusSdk.Setup
 		public List<GameObject> SpaceMenus;
 
 		public GameObject RemoteUserAvatar;
-
-
+		
 		[System.Serializable]
 		public class CavrnusSpawnableObject
 		{
@@ -197,12 +198,15 @@ namespace CavrnusSdk.Setup
 						CurrentSpaceLoadingUi.Add(Instantiate(ui, UiCanvas.transform));
 			}
 
-			CavrnusFunctionLibrary.AwaitAnySpaceConnection(spaceId => FinalizeSpaceJoin(required));
+			CavrnusFunctionLibrary.AwaitAnySpaceConnection(spaceId => FinalizeSpaceJoin(spaceId, required));
 		}
 
 		private readonly List<GameObject> currentSpaceUi = new List<GameObject>();
-		private void FinalizeSpaceJoin(bool required)
+		private void FinalizeSpaceJoin(CavrnusSpaceConnection conn, bool required)
 		{
+			if (RemoteUserAvatar != null)
+				conn.BindSpaceUsers(UserAdded, UserRemoved);
+			
 			if (required && UiCanvas == null)
 				throw new System.Exception("Error on Cavrnus Spatial Connector object: No Canvas has been specified to contain the spawned UI!");
 
@@ -213,6 +217,46 @@ namespace CavrnusSdk.Setup
 
 				foreach (var ui in SpaceMenus)
 					currentSpaceUi.Add(Instantiate(ui, UiCanvas.transform));
+			}
+		}
+
+		private Dictionary<string, GameObject> avatarInstances = new Dictionary<string, GameObject>();
+
+		//Instantiate avatars when we get a new user
+		private void UserAdded(CavrnusUser user)
+		{
+			//This list contains the player. But we don't wanna show their avatar via this system.
+			if (user.IsLocalUser)
+				return;
+
+			var avatar = Instantiate(RemoteUserAvatar, transform);
+			avatar.AddComponent<CavrnusUserFlag>().User = user;
+
+			CavrnusPropertyHelpers.ResetLiveHierarchyRootName(avatar, $"{user.ContainerId}");
+
+			foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<bool>>())
+				sync.SendMyChanges = false;
+			foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<float>>())
+				sync.SendMyChanges = false;
+			foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<Color>>())
+				sync.SendMyChanges = false;
+			foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<Vector4>>())
+				sync.SendMyChanges = false;
+			foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<CavrnusTransformData>>())
+				sync.SendMyChanges = false;
+			foreach (var sync in gameObject.GetComponentsInChildren<CavrnusValueSync<string>>())
+				sync.SendMyChanges = false;
+
+			avatarInstances[user.ContainerId] = avatar;
+		}
+
+		//Destroy them when we lose that user
+		private void UserRemoved(CavrnusUser user)
+		{
+			if (avatarInstances.ContainsKey(user.ContainerId))
+			{
+				Destroy(avatarInstances[user.ContainerId].gameObject);
+				avatarInstances.Remove(user.ContainerId);
 			}
 		}
 	}